requirements.txt
Django==1.8.5
django-haystack==2.4.1
Pillow==3.0.0
psycopg2==2.6.1
wheel==0.24.0
Whoosh==2.7.0
selenium==2.48.0

